traverse through pointer structs
If a nested struct is a pointer, still traverse through it but only if
it is already initialized. I thought maybe envstruct could help you
initialize your pointer structs if it was not initialized but I decided
against it. This is because envstruct will end up initializing all your
pointer structs even if no env value is found. This can end up causing
not nil checks against the pointer structs to be wrong because envstruct
initialized the pointers. Therefore I decided to just have envstruct
only traverse pointer structs only if it has already been initialized.

Signed-off-by: Clara Fu <fclara@vmware.com>